Libraries refactored
- Duplicates in helper functions removed and comined in madcat.helper.{h|c} and  udp_ip_port_mon.icmp_mon.helper.{h|c} .
	-> I also thought about merging the parsers, but in my opinion there are too many subtle differences in the current version.
