+++ v1_Maintenance/include/udp_ip_port_mon.helper.h
@@ -24,12 +24,12 @@ This file is part of MADCAT, the Mass Attack Detection Acceptance Tool.
     Programm erhalten haben. Wenn nicht, siehe <https://www.gnu.org/licenses/>.
 *******************************************************************************/
 /* MADCAT - Mass Attack Detecion Connection Acceptance Tool
- * ICMP monitor parser headerfile.
+ * UDP monitor library headerfile.
  *
  * Netfilter should be configured to block outgoing ICMP Destination unreachable (Port unreachable) packets, e.g.:
  *      iptables -I OUTPUT -p icmp --icmp-type destination-unreachable -j DROP
  *
- * Heiko Folkerts, BSI 2018-2019
+ * Heiko Folkerts, BSI 2018-2020
 */
 
 
@@ -37,18 +37,10 @@ This file is part of MADCAT, the Mass Attack Detection Acceptance Tool.
 #define UDP_IP_PORT_MON_HELPER_H
 
 #include "udp_ip_port_mon.h"
+#include "madcat.helper.h"
+#include "udp_ip_port_mon.icmp_mon.helper.h"
 
 //Helper Functions:
-void print_help(char* progname); //print help message
-const char* get_config_opt(lua_State* L, char* name); //Returns configuration items from LUA config file
-void get_user_ids(struct user_t* user); //adapted example code from manpage getpwnam(3)
-void print_hex(FILE* output, const unsigned char* buffer, int buffsize);
-char *print_hex_string(const unsigned char* buffer, unsigned int buffsize); //Do not forget to free!
-unsigned char* hex_dump(const void *addr, int len, const bool json);
-char* json_do(bool init_or_reset, const char* format, ...); //Reset or initialize new JSON if first arguement is true and append formated string.
-void time_str(char* buf, int buf_size);
-char *inttoa(uint32_t i_addr); //inet_ntoa e.g. converts 127.1.1.1 to 127.0.0.1. This is bad e.g. for testing.
-void* saved_buffer(void * buffer); //saves and returns address of main buffer to be freed by signal handler
-void sig_handler(int signo); //Signal Handler for gracefull shutdown
+void print_help_udp(char* progname); //print UDP help message
 
 #endif
