+++ v1_Maintenance/lib/tcp_ip_port_mon.worker.c
@@ -41,7 +41,7 @@ This file is part of MADCAT, the Mass Attack Detection Acceptance Tool.
 
 //Listner thread
 
-long int do_stuff(char* dst_addr, int dst_port, char* src_addr, int src_port, double timeout, char* data_path, int max_file_size, int s,\
+long int worker_tcp(char* dst_addr, int dst_port, char* src_addr, int src_port, double timeout, char* data_path, int max_file_size, int s,\
                   char* start_time, char* start_time_unix, FILE* confifo)
 {
         //on some systems, e.g. VMs, binding to a specific address does not work as expected.
